Fix async on vendor.js
I noticed that if you've enabled the `useAsync` option Ember apps won't run in development. This is because the vendor.js can not be loaded async (for whichever reason I don't know).

So I added an obvious check to avoid marking vendor.js as async. This shouldn't affect production where we won't have a vendor.js file.